reorganize the deprecated parts
